apigateway/requirements.txt
Flask==2.3.2
grpcio==1.57.0
grpcio-tools==1.57.0
requests==2.31.0
